% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_shock_data}
\alias{plot_shock_data}
\title{Plot Shock Data}
\usage{
plot_shock_data(
  vec,
  fps,
  shock_on_times,
  shock_off_times,
  col = "pink",
  event = "Shock",
  remove = 2,
  ymax = NULL
)
}
\arguments{
\item{vec}{Numeric array of swimming velocity data rows are cells and colummns are video frames.}

\item{fps}{Numeric value indicating frames per second.}

\item{shock_on_times}{Numeric vector of shock on times in seconds.}

\item{shock_off_times}{Numeric vector of shock off times in seconds.}
}
\description{
This function plots shock data along with average swimming velocity and
generates a bar graph showing the grand mean with confidence intervals.
}
\examples{
filename<-system.file('data/out_15000_mV_rep03.csv', package='paramecium')
dat<-read.table(filename, sep=',', header=TRUE)
vec<-lapply(unique(dat$ID), function(x){sqrt(diff(dat$X[dat$ID==x])^2+diff(dat$Y[dat$ID==x])^2)  })
vec<-do.call("rbind", vec)

logfile<-system.file('data/log_15000_mV_rep03.txt', package='paramecium')
log<-read.table(logfile, skip=1, header=TRUE, fill=TRUE, row.names = NULL)
names(log)<-c('Frame', 'Time', 'Stimuli', 'Event', 'Trial')

fps<-30  # Replace with actual frames per second

shock_on_times<-log$Frame[which(log$Stimuli=='US' & log$Event=='on')]/fps
shock_off_times<-log$Frame[which(log$Stimuli=='US' & log$Event=='off')]/fps

out<-plot_shock_data(vec, fps, shock_on_times, shock_off_times)

}
