% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_prot.R
\name{viz_protocol}
\alias{viz_protocol}
\title{Visualize Protocol Data}
\usage{
viz_protocol(
  df,
  main = "My experiment",
  exp = 0,
  colCS = "lightblue",
  colUS = "pink",
  hCS = 0.25,
  hUS = 0.125
)
}
\arguments{
\item{df}{A data frame containing protocol data.}

\item{exp}{which experiment to choose. Default is 0.}

\item{colCS}{color of the CS. Default is 'lightblue'.}

\item{colUS}{color of the US. Default is 'pink'.}
}
\description{
This function generates visualizations for the data frame containing protocol information.
}
\details{
This function takes a data frame 'df' as input and creates various visualizations
to help understand the protocol information.
}
\examples{
filename<-system.file('data/exp3.xlsx', package='paramecium')
df<-read_xlsx(path = filename)
viz_protocol(bank_df, main='My experiment', hUS=0.01)

}
\keyword{data}
\keyword{visualization}
